+++ build_runner/test/server/serve_handler_test.dart
@@ -128,7 +128,7 @@ void main() {
       node.deletedBy.add(node.id.addExtension('.post_anchor.1'));
     }
     assetGraph.add(node);
-    readerWriter.cacheStringAsset(node.id, content);
+    readerWriter.filesystem.writeAsStringSync(node.id, content);
   }
 
   test('can get handlers for a subdirectory', () async {