+++ build_test/lib/src/in_memory_writer.dart
@@ -2,4 +2,4 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'in_memory_reader.dart' show InMemoryAssetWriter, RecordingAssetWriter;
+export 'in_memory_reader.dart' show InMemoryAssetWriter;